Added additional notation to power spectrum plot
Added wrapper to matplotlib import to avoid crashing parallel engines.

The latter change is necessary because the multipendulum module imports
matplotlib.pyplot without a DISPLAY environment variable set. If matplotlib
is trying to use one of the backends built for interactivity and can't find
the display, there's a core dump.

The real fix is to pull the plotting out into a separate module, but I haven't
done that yet.